move is_error_count_healthy to its own file and added types file
