+++ src/sentry/api/endpoints/release_thresholds/types.py
@@ -1,0 +1,23 @@
+from datetime import datetime
+from typing import Any, Dict, TypedDict
+
+
+class SerializedThreshold(TypedDict):
+    date: datetime
+    environment: Dict[str, Any] | None
+    project: Dict[str, Any]
+    release: str
+    threshold_type: int
+    trigger_type: str
+    value: int
+    window_in_seconds: int
+
+
+class EnrichedThreshold(SerializedThreshold):
+    end: datetime
+    is_healthy: bool
+    key: str
+    project_slug: str
+    project_id: int
+    start: datetime
+    metric_value: int | None
